requirements.txt
pandas==2.2.2
numpy==2.1.1
requests==2.32.3
PyYAML==6.0.2
matplotlib==3.9.0
seaborn==0.13.2
